+++ Perebor/src/functions.h
@@ -1,6 +1,6 @@
 #ifndef FUNCTIONS_H
 #define FUNCTIONS_H
 
-int full_enumeration(int numbers, double maxWeight, double Weight[], double Cost[]);
+double full_enumeration(int numbers, double maxWeight, double Weight[], double Cost[]);
 
 #endif //FUNCTIONS_H
